move android related code to separate folder
