+++ Cargo.toml
@@ -33,20 +33,9 @@ env_logger = "0.11"
 wasm-bindgen-futures = "0.4"
 web-sys = "0.3.70"           # to access the DOM (to hide the loading text)
 
-[target.'cfg(target_os = "android")'.lib]
-crate-type=["lib", "cdylib"]
-
+# android:
 [target.'cfg(target_os = "android")'.dependencies]
 android_logger = { version = "0.14", default-features = false}
-egui-winit = { version = "0.30", default-features = false, features = ["clipboard","links"]}
-eframe = { version = "0.30", default-features = false, features = [
-    "accesskit",     # Make egui compatible with screen readers. NOTE: adds a lot of dependencies.
-    "default_fonts", # Embed the default egui fonts.
-    "glow",          # Use the glow rendering backend. Alternative: "wgpu".
-    "persistence",   # Enable restoring app state when restarting the app.
-    "android-native-activity", 
-] }
-
 
 [profile.release]
 opt-level = 2 # fast and small wasm
