+++ android/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "eframe_template_android"
+version = "0.1.0"
+authors = ["Emil Ernerfeldt <emil.ernerfeldt@gmail.com>"]
+edition = "2021"
+include = ["LICENSE-APACHE", "LICENSE-MIT", "**/*.rs", "Cargo.toml"]
+rust-version = "1.81"
+
+[lib]
+crate-type=["lib", "cdylib"]
+
+[dependencies]
+log = "0.4"
+eframe_template = { path = "../" }
+android_logger = { version = "0.14", default-features = false}
+egui-winit = { version = "0.30", default-features = false, features = ["clipboard","links"]}
+eframe = { version = "0.30", default-features = false, features = [
+    "accesskit",     # Make egui compatible with screen readers. NOTE: adds a lot of dependencies.
+    "default_fonts", # Embed the default egui fonts.
+    "glow",          # Use the glow rendering backend. Alternative: "wgpu".
+    "persistence",   # Enable restoring app state when restarting the app.
+    "android-native-activity", 
+] }
+
