+++ android/manifest.yaml
