+++ android/src/lib.rs
@@ -1,0 +1,22 @@
+use eframe_template::TemplateApp;
+use egui_winit::winit;
+
+#[no_mangle]
+fn android_main(app: winit::platform::android::activity::AndroidApp) {
+    std::env::set_var("RUST_BACKTRACE", "full");
+    android_logger::init_once(
+        android_logger::Config::default().with_max_level(log::LevelFilter::Info),
+    );
+
+    let native_options = eframe::NativeOptions {
+        android_app: Some(app),
+        ..Default::default()
+    };
+
+    eframe::run_native(
+        "eframe template",
+        native_options,
+        Box::new(|cc| Ok(Box::new(TemplateApp::new(cc)))),
+    )
+    .expect("Failed to start eframe context")
+}
